making the raw nhanes data filename static
the commit system takes care of saving snapshots of this data set at
specific dates if we ever need to go back and look through different
versions of it. Making the name static allows me to write code that
depends on the name of the file and not have to update it whenever we
get a new version of the file.